# FreeTypeFontGenerator with user-supplied PixmapPacker improvements