--- a/CHANGES
+++ b/CHANGES
@@ -6,7 +6,9 @@
 - FreeTypeFontGenerator can now render glyphs on the fly.
 - Attribute now implements Comparable, custom attributes might need to be updated, see: https://github.com/libgdx/libgdx/wiki/Material-and-environment#custom-attributes
 - API Change: Removed (previously deprecated) GLTexture#createTextureData/createGLHandle, Ray#getEndPoint(float), Color#tmp, Node#parent/children, VertexAttribute#Color(), Usage#Color, ModelBuilder#createFromMesh, BoundingBox#getCenter()/updateCorners()/getCorners(), Matrix4.tmp
-- Added PixmapPacker.getTextureRegions() method.
+- Added PixmapPacker.updateTextureRegions() method.
+- Added ability to pack "anonymous" pixmaps into PixmapPacker, which will appear in the generated texture but not a generated or updated TextureAtlas
+- Added PixmapPacker.packDirectToTexture() methods.
 - API Change: PixmapPacker.generateTextureAtlas(...) now returns an atlas which can be updated with subsequent calls to PixmapPacker.updateTextureAtlas(...)
 - API Change: FreeTypeFontGenerator.generateFont(...) now works with a user-provided PixmapPacker.