--- a/gdx/src/com/badlogic/gdx/graphics/VertexAttributes.java
+++ b/gdx/src/com/badlogic/gdx/graphics/VertexAttributes.java
@@ -31,8 +31,6 @@ public final class VertexAttributes implements Iterable<VertexAttribute>, Compar
 	 * @author mzechner */
 	public static final class Usage {
 		public static final int Position = 1;
-		/** @deprecated use {@link #ColorUnpacked} instead. */
-		@Deprecated public static final int Color = 2;
 		public static final int ColorUnpacked = 2;
 		public static final int ColorPacked = 4;
 		public static final int Normal = 8;