… gdx/src/com/badlogic/gdx/graphics/g3d/particles/batches/BillboardParticleBatch.java
… gdx/src/com/badlogic/gdx/graphics/g3d/shaders/DefaultShader.java
… gdx/src/com/badlogic/gdx/graphics/g3d/utils/MeshBuilder.java
… gdx/src/com/badlogic/gdx/graphics/g3d/utils/ModelBuilder.java
… gdx/src/com/badlogic/gdx/math/Frustum.java
… gdx/src/com/badlogic/gdx/math/Intersector.java
… gdx/src/com/badlogic/gdx/math/Matrix4.java
… gdx/src/com/badlogic/gdx/math/collision/BoundingBox.java
… gdx/src/com/badlogic/gdx/math/collision/Ray.java
… gdx/src/com/badlogic/gdx/scenes/scene2d/Event.java
… gdx/src/com/badlogic/gdx/scenes/scene2d/utils/SpriteDrawable.java
… gdx/src/com/badlogic/gdx/utils/JsonWriter.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/EdgeDetectionTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/FloatTextureTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/FramebufferToTextureTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/PixmapPackerTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/ProjectTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/bullet/BulletConstructor.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/bullet/FrustumCullingTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/bullet/SoftBodyTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/bullet/VehicleTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/extensions/FreeTypeAtlasTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/extensions/FreeTypePackTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/g3d/BaseG3dTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/g3d/LightsTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/g3d/ModelTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/g3d/ShaderCollectionTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/g3d/ShadowMappingTest.java
… tests/gdx-tests/src/com/badlogic/gdx/tests/utils/GdxTests.java