As discussed in #2980, I've made some changes to the PixmapPacker so that it can provide an array of TextureRegions based on its pages, and FreeTypeFontGenerator so it uses this method if the user provides their own packer as a FreeTypeFontParameter.

I've updated the PixmapPackerTest to both generate and update the TextureAtlas, to use different-sized pixmaps (and not just the same one with three different names despite creating three of them :), and to show the resulting texture atlas (change pages by hitting 0-9).

FreeTypePackTest now uses the new method.  I've also created a new FreeTypeAtlasTest which uses the less efficient (but more convenient) method of just generating the fonts directly.  I ran the tests before and after my changes on my laptop, and the times reported by those tests are comparable (but Nathan said that the performance difference would be worse on mobile).

Before my change, FreeTypePackTest
943 ms, 596 ms, 604 ms, 676 ms

After my change, FreeTypePackTest
608 ms, 666 ms, 640 ms

After my change, new FreeTypeAtlasTest
643 ms, 646 ms, 623 ms