@RobRendell 2015-04-10
I've made some further changes, some of which are in line with your suggestions :)

* `PixmapPacker` can pack Pixmaps "anonymously", through a new `PixmapPacker.pack(Pixmap)` method.  `FreeTypeFontGenerator` uses this feature, and no longer constructs names for its glyphs.
* `PixmapPacker` can pack direct to the texture (if the texture exists), via `PixmapPacker.packDirectToTexture(String, Pixmap)` and `PixmapPacker.packDirectToTexture(Pixmap)`.  This will update the texture directly if it exists and doesn't have other changes pending.  `FreeTypeFontGenerator` uses this feature too, instead of updating the texture itself.
* `PixmapPacker.updateTextureAtlas` takes a boolean parameter which directs it whether to refresh textures or not.

How does it all look to you?