@RobRendell 2015-04-10
Hum, I seem to have messed up the pull/rebase/merge across my branches... Xoppa's last commit seems to be showing up as a diff in my pull request now.  Peculiar.