@NathanSweet 2015-04-10
Not sure what's going on with the diff. Maybe you can squash your commits?