@RobRendell 2015-04-10
I'm in the process of making a fresh branch off master and cherry-picking my changes over, and I was going to create a fresh pull request off that branch.  Does that work for you?