@RobRendell 2015-04-10
Ok, fresh pull request created.  I don't know if I can cancel this one, but I'll see what I can do.