@RobRendell 2015-04-10
Ah, that would be the "Close and comment" button.